% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateBins.R
\name{aggregateBins}
\alias{aggregateBins}
\title{Combine ATAC signals across bins of specific size}
\usage{
aggregateBins(sample, resDir, sId, bin_size, chrom_lengths)
}
\arguments{
\item{sample}{Input Seurat object}

\item{resDir}{Path to the result directory}

\item{sId}{Result name}

\item{bin_size}{Size of the bin e.g. 100000 for 100Kbp}

\item{chrom_lengths}{Numeric vector of chromosome sizes, specific for genome}
}
\description{
Combine ATAC signals across bins of specific size
}
