% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMetacells.R
\name{extractMetacells}
\alias{extractMetacells}
\title{Compute meta-cells out of the main}
\usage{
extractMetacells(
  resDir,
  sId,
  sample,
  targColumn,
  metacell_content = 5,
  verbose = TRUE
)
}
\arguments{
\item{resDir}{Path to the result directory}

\item{sId}{Result name}

\item{sample}{Input Seurat object to split}

\item{targColumn}{Name of annotation column to split}

\item{metacell_content}{Amount of cells for adjustment, default n=5}

\item{verbose}{Detailed output, progress messages, default TRUE}
}
\value{
Invisibly returns NULL.
}
\description{
Compute meta-cells out of the main
}
