% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customPlots.R
\name{plotCnvBlocks}
\alias{plotCnvBlocks}
\title{Function to plot CNV blocks}
\usage{
plotCnvBlocks(resDir, infercnvObj = NULL, save = FALSE, verbose = TRUE)
}
\arguments{
\item{resDir}{Path to the result directory with input}

\item{infercnvObj}{InferCnv result object. Default: NULL
If NULL, then the object will be loaded from result directory.}

\item{save}{Set TRUE to save plot to the result directory. Default: FALSE}

\item{verbose}{Detailed output, progress messages, default TRUE}
}
\value{
Returns a list of plots
}
\description{
This function creates a plot for CNV assigned/identified subclones
}
\examples{
resPath = tempfile()
inPath = system.file("extdata", "MB183_ATAC_subset.tsv.gz",
                      package = "atacInferCnv")
sAnn = system.file("extdata", "MB183_ATAC_subset.CNV_blocks_ann_n30.txt",
                      package = "atacInferCnv")
prepareAtacInferCnvInput(inPath,sAnn,resPath, targColumn = "cnvBlock",
                        ctrlGrp = "Normal", performGA = FALSE)
runAtacInferCnv(resPath)
plotCnvBlocks(resPath)
}
