% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnalysis.R
\name{prepareAtacInferCnvInput}
\alias{prepareAtacInferCnvInput}
\title{Prepare analysis for the CNV calling from ATAC data}
\usage{
prepareAtacInferCnvInput(
  dataPath,
  annPath,
  resDir,
  sId = "sample",
  targColumn = "CellType",
  ctrlGrp = "Normal",
  ctrlObj = NULL,
  metaCells = FALSE
)
}
\arguments{
\item{dataPath}{Path to the input data in 10X format}

\item{resDir}{Path to the result directory}

\item{sId}{Result name. Default: "Sample"}

\item{targColumn}{Name of the target column in annotation. Default: "CellType"}

\item{ctrlGrp}{Name for the reference control cell type. Default: "Normal"}

\item{ctrlObj}{Seurat/Signac object to use as non-tumor control. Default: NULL}

\item{annData}{Path to annotation of the cells. Should have column}

\item{meta}{Set TRUE to use meta cells, default FALSE}
}
\description{
Prepare analysis for the CNV calling from ATAC data
}
