% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customPlots.R
\name{plotCnvBlocks}
\alias{plotCnvBlocks}
\title{Function to plot CNV blocks}
\usage{
plotCnvBlocks(resDir)
}
\arguments{
\item{resDir}{Path to the result directory with input}
}
\description{
This function creates a plot for CNV assigned/identified subclones
}
