% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnalysis.R,
%   R/prepareAnalysis.external_ref_attempt.R
\name{prepareAtacInferCnvInput}
\alias{prepareAtacInferCnvInput}
\title{Prepare analysis for the CNV calling from ATAC data}
\usage{
prepareAtacInferCnvInput(
  dataPath,
  annPath,
  resDir,
  sId = "sample",
  targColumn = "seurat_clusters",
  ctrlGrp = "Normal",
  ctrlObj = NULL,
  metaCells = FALSE
)

prepareAtacInferCnvInput(
  dataPath,
  annPath,
  resDir,
  sId = "sample",
  targColumn = "seurat_clusters",
  ctrlGrp = "Normal",
  ctrlObj = NULL,
  metaCells = FALSE
)
}
\arguments{
\item{dataPath}{Path to the input data in 10X format}

\item{resDir}{Path to the result directory}

\item{sId}{Result name}

\item{targColumn}{Column name to for usage in cells annotation.}

\item{ctrlGrp}{Name for the reference control cell type. Default: "Normal"}

\item{ctrlObj}{External Seurat object to use as non-tumor control. Default: NULL}

\item{metaCells}{Activate TRUE to use use meta cells, default FALSE}

\item{meta}{Set TRUE to use meta cells, default FALSE}

\item{annData}{Path to annotation of the cells. Not required if external control is provided.}
}
\description{
Prepare analysis for the CNV calling from ATAC data

Prepare analysis for the CNV calling from ATAC data
}
