% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareAnalysis.R
\name{prepareAtacInferCnvInput}
\alias{prepareAtacInferCnvInput}
\title{Prepare analysis for the CNV calling from ATAC data}
\usage{
prepareAtacInferCnvInput(
  dataPath,
  annPath,
  resDir,
  sId = "sample",
  targColumn = "seurat_clusters",
  ctrlGrp = "Normal",
  metaCells = FALSE
)
}
\arguments{
\item{dataPath}{Path to the input data in 10X format}

\item{resDir}{Path to the result directory}

\item{sId}{Result name}

\item{ctrlGrp}{Name for the reference control cell type. Default: "Normal"}

\item{annData}{Path to annotation of the cells}

\item{meta}{TRUE if use meta cells, default FALSE}
}
\description{
Prepare analysis for the CNV calling from ATAC data
}
