% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAtacInferCnv.R
\name{runAtacInferCnv}
\alias{runAtacInferCnv}
\title{Run InferCNV calling using generated input config}
\usage{
runAtacInferCnv(resDir, configFile = "infercnv_config.yml")
}
\arguments{
\item{resDir}{Path to the result directory with input}

\item{configFile}{Name of configuration file with InferCnv params}
}
\description{
Run InferCNV calling using generated input config
}
