% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAtacInferCnv.R
\name{runAtacInferCnv}
\alias{runAtacInferCnv}
\title{Wrapper function to run InferCNV calling}
\usage{
runAtacInferCnv(
  resDir,
  configFile = "infercnv_config.yml",
  numClusters = 1,
  chrToExclude = c("Y", "MT"),
  addDenoise = TRUE,
  clusterRefs = FALSE,
  smoothMethod = "runmeans",
  ...
)
}
\arguments{
\item{resDir}{Path to the result directory with input}

\item{configFile}{Name of configuration file with InferCnv input data}

\item{numClusters}{Number of clusters for hier. clustering. If equals one (by default)
then no clustering is performed and provided annoitation used for the formation of CNV groups.}

\item{chrToExclude}{Chromosomes to exclude. Default: Y,MT}

\item{addDenoise}{Activate denoise (InferCNV param). Deafult: TRUE}

\item{clusterRefs}{Cluster also reference (InferCNV param). Default: FALSE}

\item{smoothMethod}{Method for smoothing (InferCNV param). Default: runmeans}

\item{...}{Other parameters to provide for infercnv::run, more details in documentation of this function}
}
\value{
Invisibly returns NULL.
}
\description{
This function calls InferCNV from generated input. It has support for all
original inferCnv options.
}
\examples{
resPath = tempfile()
inPath = system.file("extdata", "MB183_ATAC_subset.tsv.gz", package = "atacInferCnv")
sAnn = system.file("extdata", "MB183_ATAC_subset.CNV_blocks_ann_n30.txt", package = "atacInferCnv" )
prepareAtacInferCnvInput(inPath,sAnn,resPath, targColumn = "cnvBlock",
                         ctrlGrp = "Normal",performGA = FALSE)
runAtacInferCnv(resPath)

}
